requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pydantic_settings
python-decouple